requirements.txt
# Name                    Version                   Build  Channel
arabic-reshaper           2.1.3                    pypi_0    pypi
brotlipy                  0.7.0           py39h2bbff1b_1003  
ca-certificates           2022.4.26            haa95532_0  
certifi                   2022.6.15        py39haa95532_0  
cffi                      1.15.0           py39h2bbff1b_1  
charset-normalizer        2.0.4              pyhd3eb1b0_0  
click                     8.0.4            py39haa95532_0  
colorama                  0.4.4              pyhd3eb1b0_0  
conda                     4.13.0           py39haa95532_0  
conda-content-trust       0.1.1              pyhd3eb1b0_0  
conda-package-handling    1.8.1            py39h8cc25b3_0  
console_shortcut          0.1.1                         4  
cryptography              36.0.0           py39h21b164f_0  
filelock                  3.7.1                    pypi_0    pypi
future                    0.18.2                   pypi_0    pypi
huggingface-hub           0.8.1                    pypi_0    pypi
idna                      3.3                pyhd3eb1b0_0  
menuinst                  1.4.18           py39h59b6b97_0  
numpy                     1.23.0                   pypi_0    pypi
oauthlib                  3.2.0                    pypi_0    pypi
openssl                   1.1.1p               h2bbff1b_0  
packaging                 21.3                     pypi_0    pypi
pandas                    1.4.3                    pypi_0    pypi
pip                       21.2.4           py39haa95532_0  
powershell_shortcut       0.0.1                         3  
pyarabic                  0.6.15                   pypi_0    pypi
pycosat                   0.6.3            py39h2bbff1b_0  
pycparser                 2.21               pyhd3eb1b0_0  
pyopenssl                 22.0.0             pyhd3eb1b0_0  
pyparsing                 3.0.9                    pypi_0    pypi
pysocks                   1.7.1            py39haa95532_0  
python                    3.9.12               h6244533_0  
python-bidi               0.4.2                    pypi_0    pypi
python-dateutil           2.8.2                    pypi_0    pypi
python-dotenv             0.20.0             pyhd8ed1ab_0    conda-forge
pytz                      2022.1                   pypi_0    pypi
pywin32                   302              py39h2bbff1b_2  
pyyaml                    6.0                      pypi_0    pypi
regex                     2022.6.2                 pypi_0    pypi
requests                  2.27.1             pyhd3eb1b0_0  
requests-oauthlib         1.3.1                    pypi_0    pypi
ruamel_yaml               0.15.100         py39h2bbff1b_0  
setuptools                61.2.0           py39haa95532_0  
six                       1.16.0             pyhd3eb1b0_1  
sqlite                    3.38.2               h2bbff1b_0  
tokenizers                0.12.1                   pypi_0    pypi
tqdm                      4.63.0             pyhd3eb1b0_0  
transformers              4.20.1                   pypi_0    pypi
tweepy                    4.10.0                   pypi_0    pypi
typing-extensions         4.3.0                    pypi_0    pypi
tzdata                    2022a                hda174b7_0  
urllib3                   1.26.8             pyhd3eb1b0_0  
vc                        14.2                 h21ff451_1  
vs2015_runtime            14.27.29016          h5e58377_2  
wheel                     0.37.1             pyhd3eb1b0_0  
win_inet_pton             1.1.0            py39haa95532_0  
wincertstore              0.2              py39haa95532_2  
yaml                      0.2.5                he774522_0  
